Add new TagContainsKeywordPredicate class and refactor predicate classes to their own package
