+++ src/main/java/educonnect/logic/commands/EditCommand.java
@@ -26,7 +26,7 @@ import educonnect.model.student.Student;
 import educonnect.model.student.StudentId;
 import educonnect.model.student.TelegramHandle;
 import educonnect.model.student.timetable.Timetable;
-import educonnect.model.tag.Tag;
+import educonnect.model.student.Tag;
 
 /**
  * Edits the details of an existing student in the address book.
