+++ src/main/java/educonnect/logic/commands/FindCommand.java
@@ -5,20 +5,20 @@ import static java.util.Objects.requireNonNull;
 import educonnect.commons.util.ToStringBuilder;
 import educonnect.logic.Messages;
 import educonnect.model.Model;
-import educonnect.model.student.NameContainsKeywordsPredicate;
+import educonnect.model.student.predicates.NameContainsKeywordsPredicate;
 
 /**
  * Finds and lists all students in address book whose name contains any of the argument keywords.
- * Keyword matching is case insensitive.
+ * Keyword matching is case-insensitive.
  */
 public class FindCommand extends Command {
 
     public static final String COMMAND_WORD = "find";
 
-    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds all students whose names contain any of "
-            + "the specified keywords (case-insensitive) and displays them as a list with index numbers.\n"
-            + "Parameters: KEYWORD [MORE_KEYWORDS]...\n"
-            + "Example: " + COMMAND_WORD + " alice bob charlie";
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds all students whose criteria matches the "
+            + "the specified keywords (case-insensitive) and displays them as a list.\n"
+            + "Parameters: [n/NAME] [s/STUDENT_ID] [h/TELEGRAM_HANDLE] [t/TAG]...\n"
+            + "Example: " + COMMAND_WORD + " n/alice t/tutorial-1";
 
     private final NameContainsKeywordsPredicate predicate;
