+++ src/main/java/educonnect/logic/parser/AddCommandParser.java
@@ -27,7 +27,7 @@ import educonnect.model.student.Student;
 import educonnect.model.student.StudentId;
 import educonnect.model.student.TelegramHandle;
 import educonnect.model.student.timetable.Timetable;
-import educonnect.model.tag.Tag;
+import educonnect.model.student.Tag;
 
 /**
  * Parses input arguments and creates a new AddCommand object
