+++ src/main/java/educonnect/logic/parser/ParserUtil.java
@@ -20,7 +20,7 @@ import educonnect.model.student.timetable.Period;
 import educonnect.model.student.timetable.Timetable;
 import educonnect.model.student.timetable.exceptions.NumberOfDaysException;
 import educonnect.model.student.timetable.exceptions.OverlapPeriodException;
-import educonnect.model.tag.Tag;
+import educonnect.model.student.Tag;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
