+++ src/main/java/educonnect/storage/JsonAdaptedStudent.java
@@ -16,7 +16,7 @@ import educonnect.model.student.Student;
 import educonnect.model.student.StudentId;
 import educonnect.model.student.TelegramHandle;
 import educonnect.model.student.timetable.Timetable;
-import educonnect.model.tag.Tag;
+import educonnect.model.student.Tag;
 
 /**
  * Jackson-friendly version of {@link Student}.
