+++ src/test/java/educonnect/logic/commands/CommandTestUtil.java
@@ -19,7 +19,7 @@ import educonnect.commons.core.index.Index;
 import educonnect.logic.commands.exceptions.CommandException;
 import educonnect.model.AddressBook;
 import educonnect.model.Model;
-import educonnect.model.student.NameContainsKeywordsPredicate;
+import educonnect.model.student.predicates.NameContainsKeywordsPredicate;
 import educonnect.model.student.Student;
 import educonnect.testutil.EditStudentDescriptorBuilder;
