+++ src/test/java/educonnect/logic/parser/AddCommandParserTest.java
@@ -54,7 +54,7 @@ import educonnect.model.student.Student;
 import educonnect.model.student.StudentId;
 import educonnect.model.student.TelegramHandle;
 import educonnect.model.student.timetable.Timetable;
-import educonnect.model.tag.Tag;
+import educonnect.model.student.Tag;
 import educonnect.testutil.StudentBuilder;
 import educonnect.testutil.TypicalTimetableAndValues;
