+++ src/test/java/educonnect/logic/parser/EditCommandParserTest.java
@@ -42,7 +42,7 @@ import educonnect.model.student.Email;
 import educonnect.model.student.Name;
 import educonnect.model.student.StudentId;
 import educonnect.model.student.TelegramHandle;
-import educonnect.model.tag.Tag;
+import educonnect.model.student.Tag;
 import educonnect.testutil.EditStudentDescriptorBuilder;
 
 public class EditCommandParserTest {
