+++ src/test/java/educonnect/model/student/NameContainsKeywordsPredicateTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import educonnect.model.student.predicates.NameContainsKeywordsPredicate;
 import org.junit.jupiter.api.Test;
 
 import educonnect.testutil.StudentBuilder;
