+++ src/test/java/educonnect/testutil/EditStudentDescriptorBuilder.java
@@ -11,7 +11,7 @@ import educonnect.model.student.Student;
 import educonnect.model.student.StudentId;
 import educonnect.model.student.TelegramHandle;
 import educonnect.model.student.timetable.Timetable;
-import educonnect.model.tag.Tag;
+import educonnect.model.student.Tag;
 
 /**
  * A utility class to help with building EditStudentDescriptor objects.
