+++ src/test/java/educonnect/testutil/StudentBuilder.java
@@ -9,7 +9,7 @@ import educonnect.model.student.Student;
 import educonnect.model.student.StudentId;
 import educonnect.model.student.TelegramHandle;
 import educonnect.model.student.timetable.Timetable;
-import educonnect.model.tag.Tag;
+import educonnect.model.student.Tag;
 import educonnect.model.util.SampleDataUtil;
 
 /**
